Create SOLUTION 1
The output for the first problem is:


Aureliso De Rosa
John Doe

What we are doing is creating an object using Javascript and using console.log to display the output on the browser console. 
First we are calling the function "getFullname" which prints the value stored in the variable "fullname" in the "prop" section.
We are then storing the function call in a new variable "var".Then using console we are calling the function using the new 
variable without getting int the object and thus the function returns the global variabke "fullname" with the value "JOHN DOE".